+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 #### Features
 
+ - Switch to Java 1.8
  - Support chunking
  - Add a databaseExists method to InfluxDB interface
  - [Issue #289] (https://github.com/influxdata/influxdb-java/issues/289) Batching enhancements: Pending asynchronous writes can be explicitly flushed via `InfluxDB.flush()`.